refactor(server): consolidate inbound rate limiter into session_ws.go
Andrew (PR #501 review): "probably should just move rate limiting to the
session_ws.go". The SEC-2 per-session token-bucket enforcement already lives
in session_ws.go IncomingLoop; only the inboundRateLimiter type and its methods
sat in a separate evr_inbound_ratelimit.go. Move them next to the loop that
enforces them and rename the test file to match. No behavior change.
